Command groups (#35)
* Feat: Add parent commands

- added subCommandHandlers
- sub commands can now be accessed/executed through parent commands

* Add parentName to subcommands

* Fix: Replace executeSubComamnd with getSubCommand

* Add "isParent" to commands having subcommands

* Feat: Update the help section for subcommands

- help.js refactored
- accessing commands using Collection instead of getFiles
- error messages added
- blocking direct call to a subcommand

* Fix: Check Admin permission for subcommands

- passing Client to each command's execute()

+++ commands/server/clearChannel.js
@@ -1,3 +1,5 @@
+const { hasRole } = require("../../utils/guildMemberHandlers");
+
 module.exports = {
   name: "clearchannel",
   permission: "*",
@@ -8,7 +10,12 @@ ${prefix}clearchannel
 Simply type the command to clear the messages of the current channel.
 Be very careful when using this command!
 \`\`\``,
-  execute: (message, args, prefix) => {
+  execute: async (bot, message, args, prefix) => {
+    const admin = await hasRole(bot, message.author.id, "Admin");
+    if (admin === false) {
+      message.reply("you do not have permission to run this command.");
+      return;
+    }
     message.channel.messages.fetch().then((res) => {
       message.channel.bulkDelete(res);
     });
